requirements.txt
asgiref==3.3.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
crispy-tailwind==0.2.0
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.8.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filter==2.4.0
django-tailwind==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
httpie==2.5.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.1.0
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
